Sprint 3: Changing all int for damage and score to double, adding stats; counter as def and Hp regeneration as vit.

+++ src/model/dungeon/scoring/Scoring.java
@@ -15,7 +15,7 @@ public class Scoring implements Serializable {
     }
 
 
-    public int killMonster(Entity entity) {
+    public double killMonster(Entity entity) {
         return entity.getValue();
     }
