+++ src/model/dungeon/tile/Breakable.java
@@ -6,17 +6,17 @@ import java.io.Serializable;
  * Abstract class that represent a breakable object
  */
 public abstract class Breakable implements Serializable {
-    private int hp;
+    private double hp;
     private boolean destroyed;
 
-    public Breakable(int hp) {
+    public Breakable(double hp) {
         this.hp = hp;
         this.destroyed = false;
     }
 
     public abstract boolean isBreakable();
 
-    public void takeDamage(int damage){
+    public void takeDamage(double damage){
         if(!destroyed) {
             hp -= damage;
             if (hp <= 0) {
